+++ packages/gatsby-theme-ui-preset/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.11...gatsby-theme-ui-preset@0.0.12) (2020-05-11)
+
+**Note:** Version bump only for package gatsby-theme-ui-preset
+
+## [0.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.10...gatsby-theme-ui-preset@0.0.11) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-theme-ui-preset
+
+## [0.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.9...gatsby-theme-ui-preset@0.0.10) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-theme-ui-preset
+
+## [0.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.8...gatsby-theme-ui-preset@0.0.9) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-theme-ui-preset
+
+## [0.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.7...gatsby-theme-ui-preset@0.0.8) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-theme-ui-preset
+
+## [0.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.6...gatsby-theme-ui-preset@0.0.7) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-theme-ui-preset
+
 ## [0.0.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-ui-preset@0.0.5...gatsby-theme-ui-preset@0.0.6) (2020-05-06)
 
 **Note:** Version bump only for package gatsby-theme-ui-preset